pyproject.toml
[tool.poetry]
name = "azure-services-tests"
version = "0.1.0"
description = ""
authors = ["Sanket Patel <sanket.patel@wellsfargo.com>"]
readme = "README.md"
packages = [{include = "azure_services_tests", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.4.4"
psycopg2-binary = "^2.9.9"
psycopg2-pool = "^1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
